requirements/requirements-prod.txt
Django==2.0.4
Pillow==5.1.0
psycopg2-binary==2.7.4
pytz==2018.3
selenium==3.11.0
gunicorn==19.7.1
django-allauth==0.35.0
raven==6.6.0
pilkit==2.0
django-imagekit==4.0.2



